refactor: setup backend structure follow industry best practices
