+++ api/partimark_app/__init__.py
@@ -1,0 +1,3 @@
+"""
+PartiMark App Package
+"""
